Improved Displacement and Coloring for Explosion Shader
The explosion shader now uses two copies of texture coords.  One copy is
stored in the vertex normals, and is the same for each vertex on a seam.
It is used for displacement mapping to prevent the mesh from separating
due to (floating point errors???).  Other set is in uv coords and used
for coloring.  It is unmodified.